+++ src/main/java/com/project/sooktoring/common/controller/AwsS3Controller.java
@@ -1,6 +1,6 @@
 package com.project.sooktoring.common.controller;
 
-import com.project.sooktoring.common.service.AwsS3Service;
+import com.project.sooktoring.common.utils.S3Uploader;
 import io.swagger.annotations.Api;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -14,19 +14,18 @@ import java.util.Map;
 @RestController
 public class AwsS3Controller {
 
-    private final AwsS3Service awsS3Service;
+    private final S3Uploader s3Uploader;
 
     //이미지 업로드
     @PostMapping("/upload")
     public String uploadImg(@RequestPart("file") MultipartFile file) {
-        return awsS3Service.uploadImg(file, "test");
+        return s3Uploader.uploadImg(file, "test");
     }
 
     //이미지 삭제
     @DeleteMapping("/delete")
     public String delete(@RequestBody Map<String, String> url) {
-        awsS3Service.deleteImg(url.get("url"));
+        s3Uploader.deleteImg(url.get("url"));
         return url.get("url"); //리턴된 url 접속 안 되는 것 확인
     }
-
 }
