+++ src/main/java/com/project/sooktoring/common/utils/ProfileUtil.java
@@ -2,6 +2,8 @@ package com.project.sooktoring.common.utils;
 
 import com.project.sooktoring.common.exception.CustomException;
 import com.project.sooktoring.profile.domain.Profile;
+import com.project.sooktoring.profile.repository.ActivityRepository;
+import com.project.sooktoring.profile.repository.CareerRepository;
 import com.project.sooktoring.profile.repository.ProfileRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -13,6 +15,8 @@ import static com.project.sooktoring.common.exception.ErrorCode.*;
 public class ProfileUtil {
 
     private final ProfileRepository profileRepository;
+    private final ActivityRepository activityRepository;
+    private final CareerRepository careerRepository;
 
     public Profile getCurrentProfile() {
         return profileRepository.findByUserId(SecurityUtil.getCurrentUserId())
@@ -22,5 +26,11 @@ public class ProfileUtil {
     public Profile getProfile(Long profileId) {
         return profileRepository.findById(profileId)
                 .orElseThrow(() -> new CustomException(NOT_FOUND_PROFILE));
+    }
+
+    public void withdraw(Long profileId) {
+        activityRepository.deleteByProfileId(profileId);
+        careerRepository.deleteByProfileId(profileId);
+        profileRepository.deleteById(profileId);
     }
 }
